refactor: rename host to gateway
